--- .prettierignore
@@ -1,2 +1,0 @@
-# package.json is formatted by package managers, so we ignore it here
-package.json
