--- .vscode/launch.json
@@ -1,39 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [{
-      "type": "node",
-      "request": "launch",
-      "name": "Debug Project",
-      // we test in `build` to make cleanup fast and easy
-      "cwd": "${workspaceFolder}/build",
-      // Replace this with your project root. If there are multiple, you can
-      // automatically run the currently visible file with: "program": ${file}"
-      "program": "${workspaceFolder}/src/cli/cli.ts",
-      // "args": ["--no-install"],
-      "outFiles": ["${workspaceFolder}/build/main/**/*.js"],
-      "skipFiles": [
-        "<node_internals>/**/*.js",
-        "${workspaceFolder}/node_modules/**/*.js"
-      ],
-      "preLaunchTask": "npm: build",
-      "stopOnEntry": true,
-      "smartStep": true,
-      "runtimeArgs": ["--nolazy"],
-      "env": {
-        "TYPESCRIPT_STARTER_REPO_URL": "${workspaceFolder}"
-      },
-      "console": "externalTerminal"
-    },
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Debug Spec",
-      "program": "${workspaceRoot}/.vscode/debug-ts.js",
-      "args": ["${file}"],
-      "skipFiles": ["<node_internals>/**/*.js"],
-      // Consider using `npm run watch` or `yarn watch` for faster debugging
-      // "preLaunchTask": "npm: build",
-      // "smartStep": true,
-      "runtimeArgs": ["--nolazy"]
-    }]
-}
