--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-  "typescript.tsdk": "node_modules/typescript/lib"
-  // "typescript.implementationsCodeLens.enabled": true
-  // "typescript.referencesCodeLens.enabled": true
-}
