--- CHANGELOG.md
@@ -1,26 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
-
-<a name="0.1.9"></a>
-## [0.1.9](https://github.com/hobochild/apollo-link-segment/compare/v0.1.7...v0.1.9) (2020-04-11)
-
-
-
-<a name="0.1.7"></a>
-## [0.1.7](https://github.com/hobochild/apollo-link-segment/compare/v0.1.5...v0.1.7) (2019-03-13)
-
-
-
-<a name="0.1.6"></a>
-## [0.1.6](https://github.com/hobochild/apollo-link-segment/compare/v0.1.4...v0.1.6) (2019-03-13)
-
-
-
-<a name="0.1.4"></a>
-## [0.1.4](https://github.com/hobochild/apollo-link-segment/compare/v0.1.2...v0.1.4) (2019-03-13)
-
-
-
-<a name="0.1.2"></a>
-## 0.1.2 (2019-03-12)
