+++ package.json
@@ -1,100 +1,38 @@
 {
   "name": "apollo-link-segment",
-  "version": "0.1.9",
+  "type": "module",
+  "source": "src/index.js",
+  "exports": {
+    "require": "./dist/index.cjs",
+    "default": "./dist/index.modern.js"
+  },
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.module.js",
+  "version": "1.0.0-0",
   "description": "Auto analytics for apollo apps",
-  "main": "build/main/index.js",
-  "typings": "build/main/index.d.ts",
+  "typings": "dist/index.d.ts",
   "repository": "https://github.com/hobochild/apollo-link-segment",
   "license": "MIT",
   "keywords": [],
   "scripts": {
-    "describe": "npm-scripts-info",
-    "build": "run-s clean && run-p build:*",
-    "build:main": "tsc -p tsconfig.json",
-    "build:module": "tsc -p tsconfig.module.json",
-    "fix": "run-s fix:*",
-    "fix:prettier": "prettier \"src/**/*.ts\" --write",
-    "fix:tslint": "tslint --fix --project .",
-    "test": "run-s build test:*",
-    "test:unit": "nyc --silent ava",
-    "watch": "run-s clean build:main && run-p \"build:main -- -w\" \"test:unit -- --watch\"",
-    "cov": "run-s build test:unit cov:html && opn coverage/index.html",
-    "cov:html": "nyc report --reporter=html",
-    "cov:send": "nyc report --reporter=lcov > coverage.lcov && codecov",
-    "cov:check": "nyc report && nyc check-coverage --lines 100 --functions 100 --branches 100",
-    "doc": "run-s doc:html && opn build/docs/index.html",
-    "doc:html": "typedoc src/ --exclude **/*.spec.ts --target ES6 --mode file --out build/docs",
-    "doc:json": "typedoc src/ --exclude **/*.spec.ts --target ES6 --mode file --json build/docs/typedoc.json",
-    "doc:publish": "gh-pages -m \"[ci skip] Updates\" -d build/docs",
-    "version": "standard-version",
-    "reset": "git clean -dfx && git reset --hard && npm i",
-    "clean": "trash build test",
-    "all": "run-s reset test cov:check doc:html",
-    "prepare-release": "run-s all version doc:publish"
-  },
-  "scripts-info": {
-    "info": "Display information about the package scripts",
-    "build": "Clean and rebuild the project",
-    "fix": "Try to automatically fix any linting problems",
-    "test": "Lint and unit test the project",
-    "watch": "Watch and rebuild the project on save, then rerun relevant tests",
-    "cov": "Rebuild, run tests, then create and open the coverage report",
-    "doc": "Generate HTML API documentation and open it in a browser",
-    "doc:json": "Generate API documentation in typedoc JSON format",
-    "version": "Bump package.json version, update CHANGELOG.md, tag release",
-    "reset": "Delete all untracked files and reset the repo to the last commit",
-    "prepare-release": "One-step: clean, build, test, publish docs, and prep a release"
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest",
+    "fmt": "prettier --write ./",
+    "fmt_check": "prettier --check ./",
+    "build": "microbundle"
   },
   "engines": {
-    "node": ">=8.9"
+    "node": ">=14"
   },
   "devDependencies": {
-    "@types/graphql": "^14.0.5",
-    "apollo-link": "^1.2.8",
-    "ava": "1.0.0-beta.7",
-    "browser-env": "^3.2.5",
-    "codecov": "^3.1.0",
-    "cz-conventional-changelog": "^2.1.0",
-    "gh-pages": "^2.0.1",
-    "graphql": "^14.1.1",
-    "graphql-tag": "^2.10.1",
-    "graphql-tools": "^4.0.4",
-    "npm-run-all": "^4.1.5",
-    "nyc": "^13.1.0",
-    "opn-cli": "^4.0.0",
-    "prettier": "^1.15.2",
-    "sinon": "^7.2.3",
-    "standard-version": "^8.0.1",
-    "trash-cli": "^1.4.0",
-    "tslint": "^5.11.0",
-    "tslint-config-prettier": "^1.17.0",
-    "tslint-immutable": "^5.0.0",
-    "typedoc": "^0.13.0",
-    "typescript": "^3.1.6"
-  },
-  "ava": {
-    "failFast": true,
-    "files": [
-      "build/main/**/*.spec.js"
-    ],
-    "sources": [
-      "build/main/**/*.js"
-    ]
-  },
-  "config": {
-    "commitizen": {
-      "path": "cz-conventional-changelog"
-    }
-  },
-  "prettier": {
-    "singleQuote": true
-  },
-  "nyc": {
-    "exclude": [
-      "**/*.spec.js"
-    ]
+    "apollo-link": "^1.2.14",
+    "graphql-tag": "^2.12.6",
+    "jest": "^28.1.1",
+    "jest-environment-jsdom": "^28.1.1",
+    "microbundle": "^0.15.0",
+    "prettier": "^2.7.1",
+    "zen-observable": "^0.8.15"
   },
   "peerDependencies": {
-    "apollo-link": "^1.2.8"
+    "apollo-link": "^1.2.14"
   }
 }
