+++ src/index.d.ts
@@ -1,0 +1,12 @@
+import { OperationTypeNode } from "graphql";
+
+type Opts = {
+  analytics?: {
+    track: (eventName: string, payload: any) => void;
+  };
+  operationWhitelist?: OperationTypeNode[];
+};
+
+// TODO update any to ApolloLink
+type link = (opts: Opts) => any;
+export default link;
