+++ src/index.js
@@ -1,0 +1,3 @@
+import link from "./lib/link";
+
+export default link;
