--- src/index.ts
@@ -1,3 +1,0 @@
-import link from './lib/link';
-
-export default link;
