--- src/types/example.d.ts
@@ -1,1 +1,0 @@
-
