--- tsconfig.module.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "target": "esnext",
-    "outDir": "build/module",
-    "module": "esnext",
-    "lib": ["es6", "dom", "esnext.asynciterable", "es7"],
-    "noImplicitAny": false
-  },
-  "exclude": ["node_modules/**"]
-}
